feat: max_initcode_size is defined relative to max_code_size

+++ system_hooks/src/call_hooks/contract_deployer_temp.rs
@@ -184,7 +184,7 @@ where
 
             // Although this can be called as a part of protocol upgrade,
             // we are checking the next invariants, just in case
-            if !system.is_contract_size_allowed(bytecode_length) {
+            if !system.is_contract_size_allowed(bytecode_length as usize) {
                 return Ok(Err(
                     "Contract deployer failure: setBytecodeDetailsEVM called with invalid bytecode(length exceeds chain-configured limit)",
                 ));
