+++ system_hooks/src/call_hooks/set_bytecode_on_address.rs
@@ -162,7 +162,7 @@ where
 
     // Although this can be called as a part of protocol upgrade,
     // we are checking the next invariants, just in case
-    if !system.is_contract_size_allowed(bytecode_length) {
+    if !system.is_contract_size_allowed(bytecode_length as usize) {
         return Ok(Err(
             "Set bytecode on address failure: called with invalid bytecode(length exceeds chain-configured limit)",
         ));
